Use an app's Fragment to allow theming the UI properly
