--- WooCommerce/src/main/res/navigation/nav_graph_api_faker.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    app:startDestination="@id/apiFakerHostFragment"
-    android:id="@+id/nav_graph_api_faker">
-    <fragment
-        android:id="@+id/apiFakerHostFragment"
-        android:name="com.woocommerce.android.apifaker.ui.ApiFakerHostFragment"
-        android:label="ApiFakerHostFragment" />
-</navigation>
